fix: Attempt to handle blink messages
